Move time mapping to mapper
