+++ remark-api/src/main/java/com/stelmashchuk/remark/api/comment/CommentMapper.kt
@@ -1,13 +1,9 @@
-package com.stelmashchuk.remark.api
+package com.stelmashchuk.remark.api.comment
 
 import com.stelmashchuk.remark.api.pojo.Comment
 import com.stelmashchuk.remark.api.repositories.FullComment
-import java.time.LocalDateTime
-import java.time.format.DateTimeFormatter
 
-class CommentMapper {
-
-  private val backendFormatter = DateTimeFormatter.ISO_ZONED_DATE_TIME
+class CommentMapper(private val commentTimeMapper: CommentTimeMapper) {
 
   fun mapCommentsFullComments(comments: List<Comment>): List<FullComment> {
     return comments.map { comment ->
@@ -17,12 +13,26 @@ class CommentMapper {
           text = comment.text,
           score = comment.score,
           user = comment.user,
-          time = LocalDateTime.parse(comment.time, backendFormatter),
+          time = commentTimeMapper.map(comment.time),
           vote = comment.vote,
           replyCount = comments.count { it.parentId == comment.id },
           isCurrentUserAuthor = false
       )
     }
+  }
+
+  fun map(comment: Comment): FullComment {
+    return FullComment(
+        id = comment.id,
+        parentId = comment.parentId,
+        text = comment.text,
+        score = comment.score,
+        user = comment.user,
+        time = commentTimeMapper.map(comment.time),
+        vote = comment.vote,
+        replyCount = 0,
+        isCurrentUserAuthor = true
+    )
   }
 }
