+++ remark-api/src/test/java/com/stelmashchuk/remark/api/comment/PostCommentUseCaseTest.kt
@@ -14,7 +14,7 @@ import io.mockk.coEvery
 import io.mockk.mockk
 import kotlinx.coroutines.runBlocking
 import org.junit.jupiter.api.Test
-import java.util.*
+import java.time.LocalDateTime
 
 internal class PostCommentUseCaseTest {
 
@@ -44,9 +44,9 @@ internal class PostCommentUseCaseTest {
       }
     }
 
-    commentStorage.add(FullComment(rootId, "", "text", 0L, mockk(), Date(), 0, 0, true))
+    commentStorage.add(FullComment(rootId, "", "text", 0L, mockk(), LocalDateTime.MAX, 0, 0, true))
 
-    val userCase = PostCommentUseCase(commentStorage, remarkService)
+    val userCase = PostCommentUseCase(commentStorage, remarkService, CommentMapper(mockk(relaxed = true)))
     userCase.postComment(CommentRoot.Comment(postUrl, rootId), newText, postUrl, siteId)
 
     commentStorage.waitForComment(newId) should idMatch(newId).and(textMatch(newText)).and(replyCountMatch(0))
